+++ CHANGES.md
@@ -1,6 +1,13 @@
 Release Notes
 ==========
 
+Version 1.0.0-beta-2
+---------------
+* Added ability to store returned values in variables or in the plugin locally
+* Added menu option to show all plugin-stored-values
+* Added action that can be called to retrieve plugin-stored-values
+* Added additional error checking
+
 Version 1.0.0-beta-1
 ---------------
 * Initial beta release